removes buffering when generating coding shreds in broadcast
Given the 32:32 erasure recovery schema, current implementation requires
exactly 32 data shreds to generate coding shreds for the batch (except
for the last erasure batch in each slot).
As a result, when serializing ledger entries to data shreds, if the
number of data shreds is not a multiple of 32, the coding shreds for the
final batch cannot be generated until there are more data shreds. This
adds latency in generating and broadcasting coding shreds.

In addition, with Merkle variants for shreds, data shreds cannot be
signed until coding shreds are also generated. As a result *both* code
and data shreds will be delayed before broadcast if we still require
exactly 32 data shreds for each batch.

This commit instead always generates and broadcast coding shreds as soon
as there any number of data shreds available. There will be
    64 - num_data_shreds
coding shreds in each batch so that erasure batches stay at size 64
(with at most 32 data shreds in each batch).
A consequence of this change is that code and data shreds indices will
no longer align as there will be more coding shreds than data shreds
(not only in the last batch in each slot but also in the intermediate
ones); however with recent patches backported to the mainnet, receiving
nodes should be able to correctly digest and process these shreds.