skips (redundant) signature verification for recovered Merkle shreds
With Merkle shreds, leader signs the Merkle root of the erasure batch
and all shreds within the same erasure batch have the same signature.
For recovered shreds, the (unique) signature is copied from shreds which
were received from turbine (or repair) and are already sig-verified.
The same signature also verifies for recovered shreds because when
reconstructing the Merkle tree for the erasure batch, we will obtain the
same Merkle root.