reduces std::mem::size_of::<gossip::CrdsData>()
Because ContactInfo.cache is huge, currently
    std::mem::size_of::<gossip::CrdsData>()
stands at 560 bytes. However only a small % of entries in gossip CRDS
table are ContactInfo so this wastes a lot of memory.

In order to reduce the size of CrdsData enum, the commit adds type
parameter for ContactInfo.cache. For the ContactInfo stored in CRDS
table we use
    ContactInfo.cache: Box<[SocketAddr; N]>
whereas outside gossip CRDS table we avoid Box<...> overhead by just
using the plain array:
    ContactInfo.cache: [SocketAddr; N]

Doing so,
  * std::mem::size_of::<gossip::CrdsData>() reduces from 560 bytes to
    176 bytes.
  * Accessing other fields of ContactInfo does not incur Box<...>
    overhead.
  * Outside gossip CRDS table, we avoid Box<...> overhead entirely.