allows inserting shreds into blockstore using a reference
Shreds generated during leader slots are concurrently inserted into
blockstore while also broadcasted to turbine. Currently shreds are
shared between the two paths using an Arc<Vec<Shred>> and sent to both
BroadcastRun::record and BroadcastRun::transmit:
https://github.com/anza-xyz/agave/blob/ddec7bdbc/turbine/src/broadcast_stage/standard_broadcast_run.rs#L224-L226
https://github.com/anza-xyz/agave/blob/ddec7bdbc/turbine/src/broadcast_stage.rs#L321
https://github.com/anza-xyz/agave/blob/ddec7bdbc/turbine/src/broadcast_stage.rs#L344

However, blockstore API for inserting shreds currently require an owned
shred anyways:
https://github.com/anza-xyz/agave/blob/ddec7bdbc/ledger/src/blockstore.rs#L1231

and we cannot get an owned shred out of Arc<Vec<Shred>> if there are
other references to the Arc.

As a result StandardBroadcastRun::insert is forced to do
Arc::unwrap_or_clone anyways without any guarantee that the other
reference (turbine broadcast path) is already dropped:
https://github.com/anza-xyz/agave/blob/b59c1aee7/turbine/src/broadcast_stage/standard_broadcast_run.rs#L370

This defeats the point of using Arc<Vec<Shred>> to begin with.

The commit updates blockstore insert-shreds API to instead work with
Cow<'a, Shred>. This allows to insert shreds into the blockstore with
either a reference or an owned value, and so Arc::unwrap_or_clone in
StandardBroadcastRun::insert can be avoided.